requirements.txt
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
mock
Pillow==8.0.1
